Changes to bring the Jcc wrapper in line with Jcc 1.0a.
Note that some information is not available from the JccTpi service
provider interface or TelephonyListener callback interface, and so
JccCall.release(int) and JccConnection.release(int) cause codes are
not properly proagated to the telephony fabric, and so will not show up
in the release events.

Similarily, redirecting information is not supplied when a connection is
dropped, and so JccConnection.getRedirectedAddress always returns null.

These shortcomings should be addressed in the JccTpi and TelephonyListener
once Jcc is finalized.